requirements-image-tools.txt
numpy==2.1.2
Pillow==11.0.0
rembg==2.0.59
onnxruntime==1.20.1
